+++ ui/app/models/allocation.js
@@ -25,6 +25,7 @@ export default Model.extend({
   name: attr('string'),
   taskGroupName: attr('string'),
   resources: fragment('resources'),
+  allocatedResources: fragment('resources'),
   jobVersion: attr('number'),
 
   modifyIndex: attr('number'),