+++ ui/app/models/sidecar-proxy.js
@@ -1,0 +1,6 @@
+import Fragment from 'ember-data-model-fragments/fragment';
+import { fragmentArray } from 'ember-data-model-fragments/attributes';
+
+export default Fragment.extend({
+  upstreams: fragmentArray('sidecar-proxy-upstream'),
+});