+++ ui/mirage/factories/allocation.js
@@ -2,6 +2,7 @@ import Ember from 'ember';
 import moment from 'moment';
 import { Factory, faker, trait } from 'ember-cli-mirage';
 import { provide, pickOne } from '../utils';
+import { generateResources } from '../common';
 
 const UUIDS = provide(100, faker.random.uuid.bind(faker.random));
 const CLIENT_STATUSES = ['pending', 'running', 'complete', 'failed', 'lost'];
@@ -62,6 +63,14 @@ export default Factory.extend({
       );
 
       allocation.update({ taskResourceIds: resources.mapBy('id') });
+    },
+  }),
+
+  withAllocatedResources: trait({
+    allocatedResources: () => {
+      return {
+        Shared: generateResources({ networks: { minPorts: 2 } }),
+      };
     },
   }),